+++ src/main/java/istjbot/task/Task.java
@@ -1,3 +1,5 @@
+package istjbot.task;
+
 import java.time.LocalDate;
 import java.util.Optional;
