+++ src/main/java/istjbot/ui/Ui.java
@@ -1,3 +1,5 @@
+package istjbot.ui;
+
 import java.util.Scanner;
 
 public class Ui {
