+++ text-ui-test/EXPECTED.TXT
@@ -1,105 +1,105 @@
 *_______________________________________________________________*
-Hello! I'm IstjBot. 
+Hello! I'm istjbot.IstjBot.
 What can I do for you?
 *_______________________________________________________________*
 *_______________________________________________________________*
-As an IstjBot, I present you the task(s) in your list:
+As an istjbot.IstjBot, I present you the task(s) in your list:
 *_______________________________________________________________*
 *_______________________________________________________________*
-As an IstjBot, I cannot find any description for your task.
+As an istjbot.IstjBot, I cannot find any description for your task.
 *_______________________________________________________________*
 *_______________________________________________________________*
-As an IstjBot, I will add this task right now. 
+As an istjbot.IstjBot, I will add this task right now.
 [T][ ] something
 Now you have 1 task in the list.
 *_______________________________________________________________*
 *_______________________________________________________________*
-As an IstjBot, I cannot add a special task with no timing attached.
+As an istjbot.IstjBot, I cannot add a special task with no timing attached.
 *_______________________________________________________________*
 *_______________________________________________________________*
-As an IstjBot, I cannot add a special task with no timing attached.
+As an istjbot.IstjBot, I cannot add a special task with no timing attached.
 *_______________________________________________________________*
 *_______________________________________________________________*
-As an IstjBot, I will add this task right now. 
+As an istjbot.IstjBot, I will add this task right now.
 [D][ ] homework (by: Jan 25 2021)
 Now you have 2 tasks in the list.
 *_______________________________________________________________*
 *_______________________________________________________________*
-As an IstjBot, I cannot add a special task with no timing attached.
+As an istjbot.IstjBot, I cannot add a special task with no timing attached.
 *_______________________________________________________________*
 *_______________________________________________________________*
-As an IstjBot, I will add this task right now. 
+As an istjbot.IstjBot, I will add this task right now.
 [E][ ] party (at: Jan 26 2021)
 Now you have 3 tasks in the list.
 *_______________________________________________________________*
 *_______________________________________________________________*
-As an IstjBot, I present you the task(s) in your list:
+As an istjbot.IstjBot, I present you the task(s) in your list:
 1. [T][ ] something
 2. [D][ ] homework (by: Jan 25 2021)
 3. [E][ ] party (at: Jan 26 2021)
 *_______________________________________________________________*
 *_______________________________________________________________*
-As an IstjBot, I don't think that is a proper index.
+As an istjbot.IstjBot, I don't think that is a proper index.
 *_______________________________________________________________*
 *_______________________________________________________________*
-As an IstjBot, I don't think that is a proper index.
+As an istjbot.IstjBot, I don't think that is a proper index.
 *_______________________________________________________________*
 *_______________________________________________________________*
-As an IstjBot, I don't think that is a proper index.
+As an istjbot.IstjBot, I don't think that is a proper index.
 *_______________________________________________________________*
 *_______________________________________________________________*
-As an IstjBot, I've marked this task as done: 
+As an istjbot.IstjBot, I've marked this task as done:
 [T][X] something
 *_______________________________________________________________*
 *_______________________________________________________________*
-As an IstjBot, I've marked this task as done: 
+As an istjbot.IstjBot, I've marked this task as done:
 [D][X] homework (by: Jan 25 2021)
 *_______________________________________________________________*
 *_______________________________________________________________*
-As an IstjBot, I've marked this task as done: 
+As an istjbot.IstjBot, I've marked this task as done:
 [E][X] party (at: Jan 26 2021)
 *_______________________________________________________________*
 *_______________________________________________________________*
-As an IstjBot, I don't think that is a proper index.
+As an istjbot.IstjBot, I don't think that is a proper index.
 *_______________________________________________________________*
 *_______________________________________________________________*
-As an IstjBot, I've unmarked this task: 
+As an istjbot.IstjBot, I've unmarked this task:
 [D][ ] homework (by: Jan 25 2021)
 *_______________________________________________________________*
 *_______________________________________________________________*
-As an IstjBot, I've unmarked this task: 
+As an istjbot.IstjBot, I've unmarked this task:
 [E][ ] party (at: Jan 26 2021)
 *_______________________________________________________________*
 *_______________________________________________________________*
-As an IstjBot, I present you the task(s) in your list:
+As an istjbot.IstjBot, I present you the task(s) in your list:
 1. [T][X] something
 2. [D][ ] homework (by: Jan 25 2021)
 3. [E][ ] party (at: Jan 26 2021)
 *_______________________________________________________________*
 *_______________________________________________________________*
-As an IstjBot, I've unmarked this task: 
+As an istjbot.IstjBot, I've unmarked this task:
 [T][ ] something
 *_______________________________________________________________*
 *_______________________________________________________________*
-As an IstjBot, I've removed this task: 
+As an istjbot.IstjBot, I've removed this task:
 [T][ ] something
 Now you have 2 tasks in the list.
 *_______________________________________________________________*
 *_______________________________________________________________*
-As an IstjBot, I've removed this task: 
+As an istjbot.IstjBot, I've removed this task:
 [E][ ] party (at: Jan 26 2021)
 Now you have 1 task in the list.
 *_______________________________________________________________*
 *_______________________________________________________________*
-As an IstjBot, I don't think that is a proper index.
+As an istjbot.IstjBot, I don't think that is a proper index.
 *_______________________________________________________________*
 *_______________________________________________________________*
-As an IstjBot, I present you the task(s) in your list:
+As an istjbot.IstjBot, I present you the task(s) in your list:
 1. [D][ ] homework (by: Jan 25 2021)
 *_______________________________________________________________*
 *_______________________________________________________________*
-As an IstjBot, I cannot understand more than bye.
+As an istjbot.IstjBot, I cannot understand more than bye.
 *_______________________________________________________________*
 *_______________________________________________________________*
-Bye, I, IstjBot, will be organizing your tasks until you come back.
+Bye, I, istjbot.IstjBot, will be organizing your tasks until you come back.
 *_______________________________________________________________*
