--- text-ui-test/data/tasks.txt
@@ -1,1 +1,0 @@
-deadline / 0 / homework / 2021-01-25
